+++ R/CreateSynData.R
@@ -283,17 +283,7 @@ WriteSynSigParams <-
 #' generated samples. Each entry is the count of mutations due to one
 #' signature in one sample.
 #'
-#' @param sig.params Parameters from \code{\link{GetSynSigParamsFromExposures}}
-#'   or another source. Should be
-#'   a matrix or data frame with one column for
-#'   each signature and the following rows:
-#' \describe{
-#' \item{prob}{The proportion of tumors with the signature.}
-#' \item{mean}{The mean(log_10(number of mutations)).}
-#' \item{stdev}{The stdev(log_10(number of mutations)).}
-#' }
-#'   The rownames need to be the column names of a signature
-#'   catalog.
+#' @inheritParams GenerateSynExposureOneSample
 #'
 #' @param num.samples Number of samples to generate
 #'